refactor(tests): consolidate shared fixture factories into @qmd-team-intent-kb/test-fixtures
Extracts makeCandidate/makeCandidateWithHash, makeMemory, makePolicy, and makeAuditEvent
from 7 duplicated fixture files into a new shared packages/test-fixtures package. Each
consumer fixture file now re-exports from the shared package and retains only its
package-specific helpers (daemon config, recording logger, JSONL writers, EvaluationContext
builder, HTTP transition body, test repo factory). Net reduction of ~338 duplicate factory
body lines across consumers.

Closes bead qmd-team-intent-kb-60h.
