+++ apps/api/package.json
@@ -26,6 +26,7 @@
     "@qmd-team-intent-kb/store": "workspace:*"
   },
   "devDependencies": {
-    "@types/better-sqlite3": "^7.6.0"
+    "@types/better-sqlite3": "^7.6.0",
+    "@qmd-team-intent-kb/test-fixtures": "workspace:*"
   }
 }
