+++ apps/api/tsconfig.json
@@ -12,6 +12,7 @@
     { "path": "../../packages/schema" },
     { "path": "../../packages/common" },
     { "path": "../../packages/store" },
-    { "path": "../../packages/policy-engine" }
+    { "path": "../../packages/policy-engine" },
+    { "path": "../../packages/test-fixtures" }
   ]
 }
