+++ apps/curator/tsconfig.json
@@ -13,6 +13,7 @@
     { "path": "../../packages/common" },
     { "path": "../../packages/store" },
     { "path": "../../packages/policy-engine" },
-    { "path": "../../packages/claude-runtime" }
+    { "path": "../../packages/claude-runtime" },
+    { "path": "../../packages/test-fixtures" }
   ]
 }
