+++ apps/edge-daemon/package.json
@@ -31,6 +31,7 @@
     "pino-pretty": "^13.0.0"
   },
   "devDependencies": {
-    "@types/better-sqlite3": "^7.6.0"
+    "@types/better-sqlite3": "^7.6.0",
+    "@qmd-team-intent-kb/test-fixtures": "workspace:*"
   }
 }
