+++ apps/edge-daemon/src/__tests__/fixtures.ts
@@ -1,5 +1,4 @@
 import { randomUUID } from 'node:crypto';
-import { MemoryCandidate, GovernancePolicy } from '@qmd-team-intent-kb/schema';
 import type { DaemonConfig, DaemonDependencies, DaemonLogger } from '../types.js';
 import {
   CandidateRepository,
@@ -9,60 +8,14 @@ import {
   ExportStateRepository,
 } from '@qmd-team-intent-kb/store';
 import type Database from 'better-sqlite3';
+import type { MemoryCandidate } from '@qmd-team-intent-kb/schema';
 
-export const NOW = '2026-01-15T10:00:00.000Z';
-export const TENANT = 'team-alpha';
-
-/** Build a valid MemoryCandidate, merging optional overrides */
-export function makeCandidate(overrides?: Record<string, unknown>): MemoryCandidate {
-  return MemoryCandidate.parse({
-    id: randomUUID(),
-    status: 'inbox',
-    source: 'claude_session',
-    content: 'Use Result<T, E> for all fallible operations in the codebase',
-    title: 'Error handling convention',
-    category: 'convention',
-    trustLevel: 'medium',
-    author: { type: 'ai', id: 'claude-1' },
-    tenantId: TENANT,
-    metadata: { filePaths: ['src/utils.ts'], tags: ['error-handling'] },
-    prePolicyFlags: {},
-    capturedAt: NOW,
-    ...overrides,
-  });
-}
-
-/** Build a valid GovernancePolicy with basic rules */
-export function makePolicy(overrides?: Record<string, unknown>): GovernancePolicy {
-  return GovernancePolicy.parse({
-    id: randomUUID(),
-    name: 'Test Policy',
-    tenantId: TENANT,
-    rules: [
-      {
-        id: 'rule-secret',
-        type: 'secret_detection',
-        action: 'reject',
-        enabled: true,
-        priority: 0,
-        parameters: {},
-      },
-      {
-        id: 'rule-length',
-        type: 'content_length',
-        action: 'reject',
-        enabled: true,
-        priority: 1,
-        parameters: { min: 10, max: 50000 },
-      },
-    ],
-    enabled: true,
-    version: 1,
-    createdAt: NOW,
-    updatedAt: NOW,
-    ...overrides,
-  });
-}
+export {
+  makeCandidate,
+  makePolicy,
+  FIXED_NOW as NOW,
+  DEFAULT_TENANT as TENANT,
+} from '@qmd-team-intent-kb/test-fixtures';
 
 /** Create all daemon dependencies from an in-memory test database */
 export function makeDeps(db: Database.Database): DaemonDependencies {
@@ -77,8 +30,9 @@ export function makeDeps(db: Database.Database): DaemonDependencies {
 
 /** Create a default test config */
 export function makeConfig(overrides?: Partial<DaemonConfig>): DaemonConfig {
+  const NOW = '2026-01-15T10:00:00.000Z';
   return {
-    tenantId: TENANT,
+    tenantId: 'team-alpha',
     pollIntervalMs: 100, // fast for tests
     maxCandidatesPerCycle: 100,
     maxSpoolFileSizeBytes: 10 * 1024 * 1024,
