+++ apps/mcp-server/package.json
@@ -30,5 +30,7 @@
     "@qmd-team-intent-kb/store": "workspace:*",
     "fast-glob": "^3.3.3"
   },
-  "devDependencies": {}
+  "devDependencies": {
+    "@qmd-team-intent-kb/test-fixtures": "workspace:*"
+  }
 }
