+++ apps/mcp-server/tsconfig.json
@@ -12,6 +12,7 @@
     { "path": "../../packages/schema" },
     { "path": "../../packages/common" },
     { "path": "../../packages/store" },
-    { "path": "../../packages/claude-runtime" }
+    { "path": "../../packages/claude-runtime" },
+    { "path": "../../packages/test-fixtures" }
   ]
 }
