+++ apps/reporting/package.json
@@ -16,6 +16,7 @@
     "@qmd-team-intent-kb/common": "workspace:*"
   },
   "devDependencies": {
-    "@types/better-sqlite3": "^7.6.0"
+    "@types/better-sqlite3": "^7.6.0",
+    "@qmd-team-intent-kb/test-fixtures": "workspace:*"
   }
 }
