+++ packages/policy-engine/src/__tests__/fixtures.ts
@@ -1,32 +1,15 @@
 import { randomUUID } from 'node:crypto';
-import { MemoryCandidate, GovernancePolicy } from '@qmd-team-intent-kb/schema';
+import { GovernancePolicy } from '@qmd-team-intent-kb/schema';
+import type { MemoryCandidate } from '@qmd-team-intent-kb/schema';
 import type { EvaluationContext } from '../types.js';
+import { FIXED_NOW, DEFAULT_TENANT } from '@qmd-team-intent-kb/test-fixtures';
 
-export const FIXED_NOW = '2026-01-15T10:00:00.000Z';
-export const DEFAULT_TENANT = 'team-alpha';
-export const DEFAULT_CONTENT = 'Use Result<T, E> for all fallible operations in the codebase';
-
-/**
- * Build a valid MemoryCandidate for policy-engine rule tests.
- * All fields are valid defaults; pass overrides to vary specific properties.
- */
-export function makeCandidate(overrides?: Record<string, unknown>): MemoryCandidate {
-  return MemoryCandidate.parse({
-    id: randomUUID(),
-    status: 'inbox',
-    source: 'claude_session',
-    content: DEFAULT_CONTENT,
-    title: 'Error handling convention',
-    category: 'convention',
-    trustLevel: 'medium',
-    author: { type: 'ai', id: 'claude-1' },
-    tenantId: DEFAULT_TENANT,
-    metadata: { filePaths: [], tags: [] },
-    prePolicyFlags: { potentialSecret: false, lowConfidence: false, duplicateSuspect: false },
-    capturedAt: FIXED_NOW,
-    ...overrides,
-  });
-}
+export {
+  makeCandidate,
+  FIXED_NOW,
+  DEFAULT_TENANT,
+  DEFAULT_CONTENT,
+} from '@qmd-team-intent-kb/test-fixtures';
 
 interface MakeContextOptions {
   /** Extra hashes for dedup-check tests */
@@ -39,6 +22,9 @@ interface MakeContextOptions {
  * Build a minimal EvaluationContext wrapping the given candidate and policy.
  * Pass `options.existingHashes` for dedup-check tests.
  * Pass `options.tenantId` for tenant-match tests.
+ *
+ * This helper is policy-engine specific and is intentionally not part of the
+ * shared test-fixtures package.
  */
 export function makeContext(
   candidate: MemoryCandidate,
