+++ packages/store/package.json
@@ -24,6 +24,7 @@
     "better-sqlite3": "^12.9.0"
   },
   "devDependencies": {
-    "@types/better-sqlite3": "^7.6.0"
+    "@types/better-sqlite3": "^7.6.0",
+    "@qmd-team-intent-kb/test-fixtures": "workspace:*"
   }
 }
