+++ packages/test-fixtures/tsconfig.json
@@ -8,10 +8,5 @@
     "declarationMap": true
   },
   "include": ["src"],
-  "references": [
-    { "path": "../../packages/schema" },
-    { "path": "../../packages/common" },
-    { "path": "../../packages/store" },
-    { "path": "../../packages/test-fixtures" }
-  ]
+  "references": [{ "path": "../schema" }, { "path": "../common" }]
 }
